integrating Interfaces

+++ Interfaces/IDeathcareFacilityOptions.cs
@@ -5,6 +5,6 @@ public interface IDeathcareFacilityOptions
 {
     public int HearseCapacity { get; set; }
     public int StorageCapacity { get; set; }
-     public int ProcessingRate { get; set; }
+     public float ProcessingRate { get; set; }
     
 }
