+++ Interfaces/IFireStationOptions.cs
@@ -5,7 +5,7 @@ public interface IFireStationOptions
 {
     public int FireEngineCapacity { get; set; }
     public int FireHelicopterCapacity { get; set; }
-    public int VehicleEfficiency { get; set; }
+    public float VehicleEfficiency { get; set; }
     public int DisasterResponseCapacity { get; set; }
 
 }
