+++ Options/WelfareOfficeOptions.cs
@@ -15,9 +15,9 @@ public class WelfareOfficeOptions :
     public int GroundPollution { get; set; }
     public int AirPollution { get; set; }
     public int NoisePollution { get; set; }
-    public int ServiceCoverageRange { get; set; }
-    public int ServiceCoverageCapacity { get; set; }
-    public int ServiceCoverageMagnitude { get; set; }
+    public float ServiceCoverageRange { get; set; }
+    public float ServiceCoverageCapacity { get; set; }
+    public float ServiceCoverageMagnitude { get; set; }
